+++ activerecord/test/cases/relation/field_ordered_values_test.rb
@@ -28,7 +28,9 @@ class FieldOrderedValuesTest < ActiveRecord::TestCase
 
     order = %w[written published proposed]
     books = Book.in_order_of(:status, order)
+    assert_equal(order, books.map(&:status))
 
+    books = Book.in_order_of("status", order)
     assert_equal(order, books.map(&:status))
   end
 
@@ -59,14 +61,18 @@ class FieldOrderedValuesTest < ActiveRecord::TestCase
 
     order = %w[hardcover paperback ebook]
     books = Book.in_order_of(:format, order)
+    assert_equal(order, books.map(&:format))
 
+    books = Book.in_order_of("format", order)
     assert_equal(order, books.map(&:format))
   end
 
   def test_in_order_of_after_regular_order
     order = [3, 4, 1]
     posts = Post.where(type: "Post").order(:type).in_order_of(:id, order)
+    assert_equal(order, posts.map(&:id))
 
+    posts = Post.where(type: "Post").order(:type).in_order_of("id", order)
     assert_equal(order, posts.map(&:id))
   end
 
@@ -78,7 +84,28 @@ class FieldOrderedValuesTest < ActiveRecord::TestCase
 
     order = ["ebook", nil, "paperback"]
     books = Book.in_order_of(:format, order)
-
     assert_equal(order, books.map(&:format))
+
+    books = Book.in_order_of("format", order)
+    assert_equal(order, books.map(&:format))
+  end
+
+  def test_in_order_of_with_associations
+    Author.destroy_all
+    Book.destroy_all
+    john = Author.create(name: "John")
+    bob = Author.create(name: "Bob")
+    anna = Author.create(name: "Anna")
+
+    john.books.create
+    bob.books.create
+    anna.books.create
+
+    order = ["Bob", "Anna", "John"]
+    books = Book.joins(:author).in_order_of("authors.name", order)
+    assert_equal(order, books.map { |book| book.author.name })
+
+    books = Book.joins(:author).in_order_of(:"authors.name", order)
+    assert_equal(order, books.map { |book| book.author.name })
   end
 end